readValue for char * no longer returns copy
Previously the caller of readValue_string was responsible for freeing the
returned buffer, which caused some big problems if mixed with string
literals in the code. Now it checks for null termination in the value
stored by the BLE_Char and adds it if it doesn't exist, reallocing if
necessary.