Implement better error handling.
When an error occurs and isError wraps the failed function call, ble.error
is set to the error and BLE_CHECK_ERROR propogates up to the Energia
sketch. This is done so we just check whether there is an error instead of
worrying about which specific error there is. It also elimates repeated
comparisons to SNP_SUCCESS.