Add R/W for raw value buffer; Use String.c_str().
Before there was only reading and writing for strings, which included the
null terminator. Reading and writing for the raw uint8_t buffer does not,
and passes the buffer length with it. Replaced String to char[]
conversions that used String.toCharArray with String.c_str, which just
returns the internal buffer.