Refactor reads and writes; finish char logging
Instead of writing by casting the value buffer to a char, int, etc., we
cast the new value to uint8_t* and copy it with memcpy in a common
function. This greatly reduces duplicate code. For logging purposes, this
refactor includes storing the endianness of the data. Integers are little
endian, but floats, buffers, and strings should be printed big endian.